fix kfac registration with enable_x64 (#64)
Enabling JAX x64 caused some operations to fall back to generic KFAC
handling because Flax-provided layer parameters remain float32 while
inputs and intermediate values may become float64. That inserts dtype
promotion operations into the traced graph.

Extend the KFAC patterns to cover the promoted float32-parameter path
and keep JaQMC-created wavefunction parameters on the same float32
convention as Flax layers. This avoids treating direct self.param
parameters as an accidental x64-only special case.

Document the dtype convention in the wavefunction and optimizer guides,
and clarify that enable_x64 affects requested arrays and computations
but does not upgrade model parameters by default.

The tests cover both jax.enable_x64=True and False across the supported
KFAC blocks.

+++ src/jaqmc/wavefunction/jastrow.py
@@ -58,7 +58,7 @@ class SimpleEEJastrow(nn.Module):
         n_up, n_down = self.nspins
 
         # Learnable decay parameters (initialized to alpha_init)
-        alpha_initializer = nn.initializers.constant(self.alpha_init)
+        alpha_initializer = nn.initializers.constant(self.alpha_init, dtype=jnp.float32)
         alpha_par = self.param("alpha_par", alpha_initializer, (1,))
         alpha_anti = self.param("alpha_anti", alpha_initializer, (1,))
